requirements.txt
streamlit==1.43.2
langchain==0.2.2
langchain-community==0.2.3
faiss-cpu==1.10.0
pypdf==5.4.0
sentence-transformers==4.0.1
transformers==4.50.0
huggingface_hub==0.29.3
google-genai
pandas